add getSentSummary & getReceiveSummary for react native

+++ ZaloPay/data/src/main/java/vn/com/vng/zalopay/data/api/response/redpacket/SentBundleListResponse.java
@@ -16,7 +16,7 @@ public class SentBundleListResponse extends BaseResponse {
     public long totalOfSentAmount;
 
     @SerializedName("totalofsentbundle")
-    public long totalOfSentBundle;
+    public int totalOfSentBundle;
 
     @SerializedName("sentbundlelist")
     public List<SentBundleResponse> bundleResponseList;
